correctly populate index columns with values from string arrays
